Moved setting dispatch to public method
