+++ internal/throttler/threshold/controller_test.go
@@ -2,6 +2,7 @@ package threshold
 
 import (
 	"context"
+	"github.com/openfga/openfga/pkg/dispatch"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -21,16 +22,16 @@ func TestShouldThrottle(t *testing.T) {
 
 	t.Run("should_respect_threshold_in_ctx", func(t *testing.T) {
 		ctx := context.Background()
-		ctx = ContextWithThrottlingThreshold(ctx, 200)
+		ctx = dispatch.ContextWithThrottlingThreshold(ctx, 200)
 		require.False(t, ShouldThrottle(ctx, 190, 100, 210))
 
-		ctx = ContextWithThrottlingThreshold(ctx, 200)
+		ctx = dispatch.ContextWithThrottlingThreshold(ctx, 200)
 		require.True(t, ShouldThrottle(ctx, 205, 100, 210))
 
-		ctx = ContextWithThrottlingThreshold(ctx, 200)
+		ctx = dispatch.ContextWithThrottlingThreshold(ctx, 200)
 		require.True(t, ShouldThrottle(ctx, 211, 100, 210))
 
-		ctx = ContextWithThrottlingThreshold(ctx, 1000)
+		ctx = dispatch.ContextWithThrottlingThreshold(ctx, 1000)
 		require.True(t, ShouldThrottle(ctx, 301, 100, 300))
 	})
 }
